Updating with signing scheme changes and wallet/transaction integration
